+++ airflow_provider_hightouch/example_dags/example_hightouch_trigger_sync.py
@@ -1,8 +1,7 @@
-from datetime import timedelta
+import datetime
 
-from airflow import DAG
-from airflow.operators.latest_only import LatestOnlyOperator
-from airflow.utils.dates import days_ago
+from airflow.providers.standard.operators.latest_only import LatestOnlyOperator
+from airflow.sdk import DAG
 
 from airflow_provider_hightouch.operators.hightouch import HightouchTriggerSyncOperator
 from airflow_provider_hightouch.sensors.hightouch import HightouchSyncRunSensor
@@ -12,12 +11,12 @@ args = {"owner": "airflow"}
 with DAG(
     dag_id="example_hightouch_operator",
     default_args=args,
-    schedule_interval="@daily",
-    start_date=days_ago(1),
-    dagrun_timeout=timedelta(minutes=5),
+    schedule="@daily",
+    start_date=datetime.datetime(2024, 1, 1),
+    dagrun_timeout=datetime.timedelta(minutes=5),
 ) as dag:
 
-    latest_only = LatestOnlyOperator(task_id="latest_only", dag=dag)
+    latest_only = LatestOnlyOperator(task_id="latest_only")
 
     # This task runs async, and doesn't poll for status or fail on error
     run_async = HightouchTriggerSyncOperator(task_id="run_async", sync_id=4)
@@ -32,7 +31,8 @@ with DAG(
     sync_sensor = HightouchSyncRunSensor(
         task_id="sync_sensor",
         sync_run_id="123456",
-        sync_id="123")
+        sync_id="123",
+    )
 
     latest_only >> sync_sensor
     sync_sensor >> run_sync